+++ src/viem/interfaces/Azrael.v0.ts
@@ -1,7 +1,11 @@
 import { packPrice, PaymentToken, prepareBatch } from '../../core';
-import { Executor, SDK } from '../executor';
+import { Executor, SDK } from '../base';
+import { SupportedSDKInterfaces } from '../deployments';
 
-export default class AzrealV0SDK extends SDK {
+export default class AzrealV0SDK<
+  ContractType extends keyof SupportedSDKInterfaces,
+  ContractVersion extends keyof SupportedSDKInterfaces[ContractType]
+> extends SDK<ContractType, ContractVersion> {
   async lend(
     nftAddress: string[],
     tokenID: string[],
