+++ src/viem/interfaces/Sylvester.v0.ts
@@ -1,7 +1,11 @@
 import { NFTStandard, packPrice, PaymentToken, prepareBatch } from '../../core';
-import { Executor, SDK } from '../executor';
+import { Executor, SDK } from '../base';
+import { SupportedSDKInterfaces } from '../deployments';
 
-export default class SylvesterV0SDK extends SDK {
+export default class SylvesterV0SDK<
+  ContractType extends keyof SupportedSDKInterfaces,
+  ContractVersion extends keyof SupportedSDKInterfaces[ContractType]
+> extends SDK<ContractType, ContractVersion> {
   async lend(
     nftStandard: NFTStandard[],
     nftAddress: string[],
